kmeans test result/mlp_player_classifier/requirements.txt
pandas
numpy
pyyaml
scikit-learn
torch
onnx
onnxruntime
tqdm
